test(react-query/QueryResetErrorBoundary): relocate 'issue-9728' test and migrate to fake timers
